\name{abakaliki}
\alias{abakaliki}
\alias{abakaliki.opts}
\docType{data}
\title{ Smallpox epidemic in Abakaliki, Nigeria }
\description{
  Use MCMC to estimate parameters in the smallpox epidemic
  of Abakaliki, Nigeria also treated in the article by ONeill and
  Roberts.
}
\usage{data(abakaliki)}
\source{
  ONeill, P. D. and Roberts, G. O. (1999). Bayesian inference
  for partially observed stochastic epidemics. 
  J. R. Statist. Soc. A 162, 121--129.
}

\examples{
\dontrun{data(abakaliki)}
\dontrun{seir(abakaliki,abakaliki.opts)}
}
\keyword{datasets}
