\name{oneill}
\alias{oneill}
\alias{oneill.opts}
\docType{data}
\title{ Simulated data of 5 recovery times from O'Neill et. al}
\description{
  Use MCMC to estimate parameters in the 5 point epidemic
  in the 5 removal times epidemic (p.126) of the article by O'Neill
  and Roberts.
}
\usage{data(oneill)}
\details{
  n.a.
}
\source{
  O'Neill, P. D. and Roberts, G. O. (1999). Bayesian inference
  for partially observed stochastic epidemics. 
  J. R. Statist. Soc. A 162, 121--129.
}

\examples{
\dontrun{data(oneill)}
\dontrun{seir(oneill,oneill.opts)}
}
\keyword{datasets}
